refactor: split shared fuzz infrastructure into output_fuzz_common
Pull the truly shared fuzz machinery (`GraphCase`, the proptest
`acyclic_graph_case_strategy`, case materialization, fixture / case-spec
helpers, `bundler_options_for_case`, `collect_chunk_info`,
`run_deterministic_check`, REPL URL builder) into a new `output_fuzz_common`
library crate.

`acyclic_output_fuzz` keeps everything cycle- and acyclic-specific:
`graph_cycle_checker`, `run_case`, `dynamic_hub_case_strategy`, the JS-syntax
and entry-export validators, the output dependency-graph builder, the
CJS-cycle helpers (`is_immediate_cycle`, `cycle_involves_cjs`,
`chunk_contains_cjs_module`), the failure-message formatter, and the
`cjs_cycle_tests` unit-test module. It now depends on `output_fuzz_common`
and the `generate_fixture` binary calls into the common crate.

Add a new `deterministic_output_fuzz` test crate that hosts the
`deterministic_output` property test, also depending on `output_fuzz_common`.
