+++ deterministic_output_fuzz/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "acyclic_output_fuzz"
+name = "deterministic_output_fuzz"
 version = "0.1.0"
 edition = "2024"
 publish = false
@@ -7,14 +7,7 @@ publish = false
 [lib]
 path = "lib.rs"
 
-[[bin]]
-name = "generate_fixture"
-path = "generate_fixture.rs"
-
 [dependencies]
 output_fuzz_common = { path = "../output_fuzz_common" }
 proptest = { workspace = true }
-rolldown = { workspace = true }
-rolldown_common = { workspace = true }
 tokio = { workspace = true }
-oxc = { workspace = true }
